# internal/api/testdata/openapi.golden.json
{
  "components": {
    "schemas": {
...
      "CAAuthority": {
        "properties": {
          "certificate_pem": {
            "type": "string"
          },
          "common_name": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "extended_key_usages": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "max_path_len": {
            "type": "integer"
          },
          "not_after": {
            "format": "date-time",
            "type": "string"
          },
          "parent_id": {
            "format": "uuid",
            "type": "string"
          },
          "permitted_dns_names": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "serial": {
            "type": "string"
          },
          "signer_handle": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "common_name",
          "kind",
          "status",
          "certificate_pem",
          "signer_handle",
          "serial",
          "max_path_len",
          "created_at"
        ],
        "type": "object"
      },
      "CAAuthorityList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CAAuthority"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "CACeremonyStartRequest": {
        "properties": {
          "operation": {
            "enum": [
              "create_root",
              "create_intermediate"
            ],
            "type": "string"
          },
          "parent_id": {
            "format": "uuid",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/CASpec"
          },
          "threshold": {
            "type": "integer"
          }
        },
        "required": [
          "operation",
          "threshold",
          "spec"
        ],
        "type": "object"
      },
      "CACreateIntermediateRequest": {
        "properties": {
          "ceremony_id": {
            "format": "uuid",
            "type": "string"
          },
          "parent_id": {
            "format": "uuid",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/CASpec"
          }
        },
        "required": [
          "ceremony_id",
          "parent_id",
          "spec"
        ],
        "type": "object"
      },
      "CACreateRootRequest": {
        "properties": {
          "ceremony_id": {
            "format": "uuid",
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/CASpec"
          }
        },
        "required": [
          "ceremony_id",
          "spec"
        ],
        "type": "object"
      },
...
      "CAIssueLeafRequest": {
        "properties": {
          "csr_pem": {
            "type": "string"
          },
          "ttl_seconds": {
            "type": "integer"
          }
        },
        "required": [
          "csr_pem"
        ],
        "type": "object"
      },
...
      "CAIssuedLeaf": {
        "properties": {
          "certificate_pem": {
            "type": "string"
          },
          "not_after": {
            "format": "date-time",
            "type": "string"
          },
          "serial": {
            "type": "string"
          }
        },
        "required": [
          "certificate_pem",
          "serial",
          "not_after"
        ],
        "type": "object"
      },
      "CAKeyCeremony": {
        "properties": {
          "approvals": {
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "opener": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "type": "string"
          },
          "threshold": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "purpose",
          "threshold",
          "status",
          "approvals",
          "created_at"
        ],
        "type": "object"
      },
      "CASpec": {
        "properties": {
          "common_name": {
            "type": "string"
          },
          "extended_key_usages": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "max_path_len": {
            "type": "integer"
          },
          "permitted_dns_domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "signature_algorithm": {
            "type": "string"
          },
          "ttl_seconds": {
            "type": "integer"
          }
        },
        "required": [
          "common_name"
        ],
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "trstctl API token",
        "description": "Hashed API token resolved to a tenant-scoped principal with named trstctl permissions.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "description": "Verified browser session from OIDC, SAML, or LDAP; mutating requests also require the double-submit CSRF token.",
        "in": "cookie",
        "name": "trstctl_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Resource-oriented REST API for trstctl. Mutations require an Idempotency-Key; errors are RFC 7807 problem+json; lists use cursor pagination.",
    "title": "trstctl API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ca/authorities": {
      "get": {
        "operationId": "listCAAuthorities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAAuthorityList"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "List served CA authorities",
        "x-trstctl-permission": "issuers:read"
      }
    },
    "/api/v1/ca/authorities/intermediates": {
      "post": {
        "operationId": "createIntermediateCA",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CACreateIntermediateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAAuthority"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create a signer-backed intermediate CA after ceremony quorum",
        "x-trstctl-permission": "issuers:write"
      }
    },
    "/api/v1/ca/authorities/roots": {
      "post": {
        "operationId": "createRootCA",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CACreateRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAAuthority"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create a signer-backed root CA after ceremony quorum",
        "x-trstctl-permission": "issuers:write"
      }
    },
...
    "/api/v1/ca/authorities/{id}/issue": {
      "post": {
        "operationId": "issueHierarchyLeaf",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CAIssueLeafRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAIssuedLeaf"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Issue a leaf certificate from a served CA authority",
        "x-trstctl-permission": "certs:issue"
      }
    },
    "/api/v1/ca/ceremonies": {
      "post": {
        "operationId": "createCACeremony",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CACeremonyStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAKeyCeremony"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Start an m-of-n CA key ceremony",
        "x-trstctl-permission": "issuers:write"
      }
    },
    "/api/v1/ca/ceremonies/{id}": {
      "get": {
        "operationId": "getCACeremony",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAKeyCeremony"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Get a CA key ceremony",
        "x-trstctl-permission": "issuers:read"
      }
    },
    "/api/v1/ca/ceremonies/{id}/approvals": {
      "post": {
        "operationId": "approveCACeremony",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CAKeyCeremony"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Approve a CA key ceremony",
        "x-trstctl-permission": "issuers:write"
      }
    },
...
  }
}